+++ noodles-util/src/alignment.rs
@@ -5,3 +5,6 @@ pub mod r#async;
 
 pub mod io;
 pub mod iter;
+mod record;
+
+pub use self::record::Record;